+++ packages/web3-eth-ens/src/ens.ts
@@ -166,6 +166,26 @@ export class ENS extends Web3Context<EthExecutionAPI & Web3NetAPI> {
 	}
 
 	/**
+	 * ERC-634 - Returns the text content stored in the resolver for the specified key.
+	 * @param ENSName - The ENS name to resolve
+	 * @param key - The key to resolve https://github.com/ethereum/ercs/blob/master/ERCS/erc-634.md#global-keys
+	 * @returns - The value content stored in the resolver for the specified key
+	 */
+	public async getText(ENSName: string, key: string): Promise<string> {
+		return this._resolver.getText(ENSName, key);
+	}
+
+
+	/**
+	 * Resolves the name of an ENS node.
+	 * @param ENSName - The node to resolve
+	 * @returns - The name
+	 */
+	public async getName(ENSName: string): Promise<string> {
+		return this._resolver.getName(ENSName);
+	}
+
+	/**
 	 * Returns the X and Y coordinates of the curve point for the public key.
 	 * @param ENSName - The ENS name
 	 * @returns - The X and Y coordinates of the curve point for the public key