+++ packages/web3-eth-ens/test/unit/resolver.test.ts
@@ -211,6 +211,60 @@ describe('resolver', () => {
 		});
 	});
 
+	describe('text', () => {
+		it('getText', async () => {
+			const supportsInterfaceMock = jest
+				.spyOn(contract.methods, 'supportsInterface')
+				.mockReturnValue({
+					call: async () => Promise.resolve(true),
+				} as unknown as NonPayableMethodObject<any, any>);
+
+				const textMock = jest.spyOn(contract.methods, 'text').mockReturnValue({
+					call: jest.fn(),
+				} as unknown as NonPayableMethodObject<any, any>);
+
+				jest.spyOn(registry, 'getResolver').mockImplementation(async () => {
+					return new Promise(resolve => {
+						resolve(contract);
+					});
+				});
+	
+				await resolver.getText(ENS_NAME, "key");
+				expect(supportsInterfaceMock).toHaveBeenCalledWith(
+					interfaceIds[methodsInInterface.text],
+				);
+				expect(textMock).toHaveBeenCalledWith(namehash(ENS_NAME), "key");
+	})
+})
+
+	describe('name', () => {
+		it('getName', async () => {
+			const supportsInterfaceMock = jest
+				.spyOn(contract.methods, 'supportsInterface')
+				.mockReturnValue({
+					call: async () => Promise.resolve(true),
+				} as unknown as NonPayableMethodObject<any, any>);
+
+				const nameMock = jest.spyOn(contract.methods, 'name').mockReturnValue({
+					call: jest.fn(),
+				} as unknown as NonPayableMethodObject<any, any>);
+
+				jest.spyOn(registry, 'getResolver').mockImplementation(async () => {
+					return new Promise(resolve => {
+						resolve(contract);
+					});
+				});
+	
+				await resolver.getName(ENS_NAME);
+				expect(supportsInterfaceMock).toHaveBeenCalledWith(
+					interfaceIds[methodsInInterface.name],
+				);
+		
+				expect(nameMock).toHaveBeenCalledWith(namehash(ENS_NAME));
+		})
+	})
+	
+
 	describe('supportsInterface', () => {
 		it('check supportsInterface for non strict hex id', async () => {
 			const interfaceId = 'setAddr';